+++ wss-20211221/core/Models/DescribePackageDeductionsResponseBody.cs
@@ -201,7 +201,7 @@ namespace AlibabaCloud.SDK.Wss20211221.Models
         public int? PageNum { get; set; }
 
         /// <summary>
-        /// <para>The number of entries per page in Settings for paging.</para>
+        /// <para>The number of entries per page in a paged query. Settings the number of rows per page for paging.</para>
         /// 
         /// <b>Example:</b>
         /// <para>100</para>